+++ data/src/main/java/dk/itu/data/services/HeightCurveService.java
@@ -31,7 +31,7 @@ public class HeightCurveService {
         }
     }
     
-    public HeightCurveElement getHeightCurveForPoint(double lon, double lat) {
+    public HeightCurveElement getHeightCurveForPoint(float lon, float lat) {
         synchronized (this.repository) {
             return this.repository.getHeightCurveForPoint(lon, lat);
         }
